+++ hub/constants.py
@@ -1,1 +1,2 @@
 BYTE_PADDING = b"\0"
+MB = 1024 * 1024
