+++ hub/core/storage/__init__.py
@@ -1,3 +1,3 @@
-from hub.core.storage.s3.s3 import S3Provider  # type: ignore
+from hub.core.storage.s3 import S3Provider  # type: ignore
 from hub.core.storage.memory import MemoryProvider
 from hub.core.storage.local import LocalProvider
