+++ hub/core/storage/provider.py
@@ -9,7 +9,6 @@ class StorageProvider(ABC, MutableMapping):
     """An abstract base class for implementing a provider.
 
     To add a new provider using Provider, create a subclass and implement all 5 abstract methods below.
-    Alternatively, you can inherit from MappedProvider and have a simpler implementation.
     """
 
     @abstractmethod
