+++ requirements/common.txt
@@ -1,4 +1,3 @@
 numpy
 boto3
 boto3-stubs[essential]
-fsspec
